create starter and finished files for video 85
